[ios, macos] Enabled example insertion for macOS
The example code insertion script is now platform-agnostic.
